# Update gather and scatter to not use Argument Encoder