--- a/mlx/backend/metal/utils.h
+++ b/mlx/backend/metal/utils.h
@@ -10,20 +10,6 @@ namespace mlx::core {
 namespace {
 
 void set_array_buffer(
-    MTL::ComputeCommandEncoder* compute_encoder,
-    MTL::ArgumentEncoder* enc,
-    const array& a,
-    int idx) {
-  auto a_buf = static_cast<const MTL::Buffer*>(a.buffer().ptr());
-  auto offset = a.data<char>() -
-      static_cast<char*>(const_cast<MTL::Buffer*>(a_buf)->contents());
-  enc->setBuffer(a_buf, offset, idx);
-  // MTL::Resource usage through argument buffer needs to be explicitly
-  // flagged to enable hazard tracking
-  compute_encoder->useResource(a_buf, MTL::ResourceUsageRead);
-}
-
-void set_array_buffer(
     MTL::ComputeCommandEncoder* enc,
     const array& a,
     int idx) {