## Proposed changes

Virtual Apple silicon devices do not have support for argument encoders, which is a challenge for the CI pipeline. I just adjusted our gather and scatter kernels to not rely on  argument encoders while leaving the core function the same. No noticeable regression in performance was observed on  M3 Max 

Before:
```
====================
X (100, 64), Indices (1000000,)
MLX: 2.767ms
PyTorch: 10.588ms
====================
X (100, 1024), Indices (100000,)
MLX: 2.901ms
PyTorch: 16.767ms
====================
X (4, 1000000), Indices ()
MLX: 0.250ms
PyTorch: 0.622ms
```

After:
```
====================
X (100, 64), Indices (1000000,)
MLX: 2.214ms
PyTorch: 10.772ms
====================
X (100, 1024), Indices (100000,)
MLX: 2.893ms
PyTorch: 17.314ms
====================
X (4, 1000000), Indices ()
MLX: 0.211ms
PyTorch: 0.234ms
```

## Checklist

Put an `x` in the boxes that apply.

- [x] I have read the [CONTRIBUTING](https://github.com/ml-explore/mlx/blob/main/CONTRIBUTING.md) document
- [x] I have run `pre-commit run --all-files` to format my code / installed pre-commit prior to committing changes
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have updated the necessary documentation (if needed)