@jagrit06 2024-02-14
I don’t think we’d call it a bug - you can check the MTLDevice object to see if it has support for argument encoders and that call comes out false for virtual devices 

I suppose it has to with packing pointers in a buffer and making that redirection available requires something more (there are also some hardware mechanics involved for some versions / things you are able do with argument encoders)